TextStyles/MainPageCards(incomplete)
Created and defined all the text styles. The UI is currently being update with the new styles according to Material3 guidelines

The cards have been created separately for now, but a generic card is being tested. For now it is working properly, however its exact use case is to be determined once more of the cards that are required are designed and coded in. It is likely that this new widget type will solve most of the use cases that are required.

Clean up of the code that was pushed last time and now has been deprecated needs to be done once these use cases are finalized and the design of the application is complete.