Cards/Headers/Colors/other widgets
Defined the colours used in the application.

Changed the implementation of the headers and made them more generic
(this still needs more working)

Added a new header widget for the course overview page headers.

Started working on the progress indicator widget to display the progress of the students

Two cards in the the cards.dart file are incomplete as they are not performing the way they should be. This should be looked at further down the line.

The old cards.dart has been moved to the deprecated file to indicate the widgets created within it are no longer being used in the application.